[TID] Start adding implementation to UI.  Add a basic config system.
Have the UI hold a copy of the rectangles passed to it.
Return proper handles when creating UI items.
Build the UI tree as items are added.
Create a config cpp file.
Create a uRect to hold the games screen coords.
Change rectangles to use width/height instead of bottom/right coord.